requirements.txt
# ==============================================================================
# Modular Weapon Detection Pipeline
# Python 3.10+, CUDA 11.8+
# ==============================================================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Object Detection
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0

# Scientific
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Configuration
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Tracking
lap>=0.4.0
cython>=0.29.0
filterpy>=1.4.5

# Analysis
thop>=0.1.1
ptflops>=0.7.0

# Utilities
tqdm>=4.65.0
Pillow>=9.5.0
pytz>=2023.3

# Notebooks
jupyter>=1.0.0
